requirements.txt
ultralytics==8.3.15
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0.76
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
pyyaml>=6.0.1
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
streamlit>=1.33.0
matplotlib>=3.8.0

